% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{forecast}
\alias{forecast}
\title{Generic forecast function}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Passed to methods.}
}
\value{
The value returned by the method dispatched on \code{object}; for
\code{"dynamic_fit"} objects a \code{"dynamic_forecast"} object (see
\code{\link[=forecast.dynamic_fit]{forecast.dynamic_fit()}}).
}
\description{
Generic for extracting forecasts from a fitted model. See
\code{\link[=forecast.dynamic_fit]{forecast.dynamic_fit()}} for the method for \code{"dynamic_fit"} objects.
}
